validate and fix mess login

+++ mvc/views/Category.php
@@ -30,8 +30,8 @@
                         <td><?php echo $category['name']; ?></td>
                         <td><?php echo $category['description']; ?></td>
                         <td>
-                            <a href="./Category/update/<?php echo $category['id'];?>" class="btn btn-primary">Edit</a>
-                            <a href="./Category/delete/<?php echo $category['id'];?>" class="btn btn-danger">Delete</a>
+                            <a href="./Category/update/<?php echo $category['id']; ?>" class="btn btn-primary">Edit</a>
+                            <a href="./Category/delete/<?php echo $category['id']; ?>" class="btn btn-danger" < href="deletelink" onclick="return confirm('Bạn chắc chắn muốn xóa?')">Delete</a>
                         </td>
                     </tr>
                 <?php } ?>
@@ -39,8 +39,5 @@
         </table>
     </div>
 </body>
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
 
 </html>
